+++ packages/blockchain-info-components/src/Tooltip/TooltipIcon.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+import { Icon } from '../Icons'
+
+const TooltipIcon = styled(Icon)`
+  display: inline-flex;
+  margin-left: 5px;
+`
+
+export default TooltipIcon
