+++ packages/blockchain-info-components/src/index.js
@@ -37,4 +37,4 @@ export { Table, TableCell, TableHeader, TableRow } from './Tables'
 export { Text, TextGroup } from './Text'
 export { Themes } from './Themes'
 export { Toast } from './Toasts'
-export { default as Tooltip, TooltipRebuild } from './Tooltip'
+export { default as Tooltip, TooltipHost, TooltipIcon } from './Tooltip'
