+++ packages/blockchain-wallet-v4-frontend/src/modals/CoinifyExchangeData/ISignThis/index.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actions } from 'data'
 import { equals, path, prop } from 'ramda'
-import { Button, Text, Tooltip, Icon } from 'blockchain-info-components'
+import { Button, Text, TooltipHost, Icon } from 'blockchain-info-components'
 import { FormattedMessage } from 'react-intl'
 
 import Helper from 'components/BuySell/FAQ'
@@ -291,11 +291,9 @@ class ISignThisContainer extends Component {
                         values={{ fiatValue: getExpiredFiatValues(q) }}
                       />
                     </QuoteExpiredText>
-                    <Icon
-                      data-tip
-                      data-for='isx.expiredtooltip'
-                      name='question-in-circle'
-                    />
+                    <TooltipHost id='isx.expiredtooltip'>
+                      <Icon name='question-in-circle' />
+                    </TooltipHost>
                   </Fragment>
                 )
               } else {
@@ -312,12 +310,6 @@ class ISignThisContainer extends Component {
               }
             })
             .getOrElse(null)}
-          <Tooltip id='isx.expiredtooltip'>
-            <FormattedMessage
-              id='scenes.buysell.coinify.isx.expiredtooltip'
-              defaultMessage='This is an estimated quote. The original quote for this trade expired. The exact amount of bitcoin received depends on when the payment is received.'
-            />
-          </Tooltip>
         </TimerContainer>
         <ISXContainer>
           <IframeWrapper>
