+++ packages/blockchain-wallet-v4-frontend/src/modals/CoinifyExchangeData/index.js
@@ -10,12 +10,7 @@ import { selectors } from 'data'
 import Create from './Create'
 import Confirm from './Confirm'
 import ISignThis from './ISignThis'
-import {
-  ModalHeader,
-  ModalBody,
-  Text,
-  TooltipRebuild
-} from 'blockchain-info-components'
+import { ModalHeader, ModalBody, Text } from 'blockchain-info-components'
 import { FormattedMessage } from 'react-intl'
 import { getData } from './selectors'
 import { path } from 'ramda'
@@ -59,10 +54,6 @@ class CoinifyExchangeData extends React.PureComponent {
     /* eslint-disable */
     this.setState({ show: true })
     /* eslint-enable */
-  }
-
-  componentDidUpdate () {
-    TooltipRebuild()
   }
 
   handleClose () {
