+++ packages/blockchain-wallet-v4-frontend/src/modals/RequestBtc/FirstStep/template.success.js
@@ -6,7 +6,13 @@ import { Field, reduxForm } from 'redux-form'
 
 import { required } from 'services/FormHelper'
 import { invalidAmountMin, invalidAmountMax } from './validation'
-import { Button, Separator, Text, Icon } from 'blockchain-info-components'
+import {
+  Button,
+  Separator,
+  Text,
+  TooltipIcon,
+  TooltipHost
+} from 'blockchain-info-components'
 import {
   FiatConvertor,
   Form,
@@ -39,10 +45,6 @@ const QRText = styled(Text)`
 const CoinSelector = styled(FormGroup)`
   width: 50%;
 `
-const TooltipIcon = styled(Icon)`
-  display: inline-flex;
-  margin-left: 5px;
-`
 
 const FirstStep = props => {
   const {
@@ -74,11 +76,9 @@ const FirstStep = props => {
                 id='modals.requestbitcoin.firststep.share'
                 defaultMessage='Copy & Share Address: '
               />
-              <TooltipIcon
-                name='question-in-circle'
-                data-tip
-                data-for='reqBitcoinShare'
-              />
+              <TooltipHost id='reqBitcoinShare'>
+                <TooltipIcon name='question-in-circle' />
+              </TooltipHost>
             </div>
             <QRText
               size='14px'
