+++ packages/blockchain-wallet-v4-frontend/src/scenes/Exchange/Shapeshift/SecondStep/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { TooltipRebuild } from 'blockchain-info-components'
 
 import { actions } from 'data'
 import { getData } from './selectors'
@@ -12,10 +11,6 @@ import Success from './template.success'
 class SecondStepContainer extends React.Component {
   componentDidMount () {
     this.props.actions.secondStepInitialized()
-  }
-
-  componentDidUpdate () {
-    TooltipRebuild()
   }
 
   render () {
