+++ packages/blockchain-wallet-v4-frontend/src/scenes/Home/ActivityList/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
-import { TooltipRebuild } from 'blockchain-info-components'
 
 import { actions, selectors } from 'data'
 import { getData } from './selectors'
@@ -18,10 +17,6 @@ class ActivityListContainer extends React.PureComponent {
 
   componentDidMount () {
     this.props.actions.initialized()
-  }
-
-  componentDidUpdate () {
-    TooltipRebuild()
   }
 
   handleRequest () {
