+++ packages/blockchain-wallet-v4-frontend/src/scenes/Transactions/Bch/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
 import Content from './Content'
-import TxFeedTooltips from '../../../components/TxFeedTooltips'
 
 const Wrapper = styled.div`
   width: 100%;
@@ -11,7 +10,6 @@ const BchTransactionsContainer = () => {
   return (
     <Wrapper>
       <Content />
-      <TxFeedTooltips />
     </Wrapper>
   )
 }
