+++ packages/blockchain-wallet-v4-frontend/src/scenes/Transactions/Eth/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import styled from 'styled-components'
 
 import Content from './Content'
-import TxFeedTooltips from '../../../components/TxFeedTooltips'
 
 const Wrapper = styled.div`
   width: 100%;
@@ -12,7 +11,6 @@ const EtherTransactionsContainer = () => {
   return (
     <Wrapper>
       <Content />
-      <TxFeedTooltips />
     </Wrapper>
   )
 }
